Added Piano Lamp Stand design
A 3D-printable base for the Piano-Lamp "Rondo" (part no. L430X) from JAHN Pianoteile.

The stand allows the use of the lamp on a digital grand piano (Kawai NV10). The base is designed in a way that it can be printed without support material. Internally it has two chambers that are connected and can be filled for example with quartz sand to make the base heavier. The base is sealed by gluing the small lid into the lower opening with epoxy resin. I also used self-adhesive red felt cut into a circular shape and glued to the bottom of the piece to protect the surface of the grand piano.